# Bloat binary size for fantastic cold start performance