--- a/common/src/lib.rs
+++ b/common/src/lib.rs
@@ -1,0 +1,1 @@
+pub mod ngram;