--- a/common/src/lib.rs
+++ b/common/src/lib.rs
@@ -1,1 +1,2 @@
+pub mod fraction;
 pub mod ngram;