--- a/language-models/ar/build.rs
+++ b/language-models/ar/build.rs
@@ -1,4 +1,4 @@
 fn main() {
     model_builder::build_frequencies("models");
-    model_builder::generate_statics("../models", "src");
+    model_builder::generate_statics("models", "src");
 }