--- a/language-models/cs/build.rs
+++ b/language-models/cs/build.rs
@@ -1,0 +1,4 @@
+fn main() {
+    model_builder::build_frequencies("models");
+    model_builder::generate_statics("../models", "src");
+}