--- a/language-models/cy/src/lib.rs
+++ b/language-models/cy/src/lib.rs
@@ -19,3 +19,6 @@ use include_dir::{include_dir, Dir};
 pub const WELSH_MODELS_DIRECTORY: Dir = include_dir!("$CARGO_MANIFEST_DIR/models");
 
 pub const WELSH_TESTDATA_DIRECTORY: Dir = include_dir!("$CARGO_MANIFEST_DIR/testdata");
+
+mod generated_statics;
+pub use generated_statics::NGRAMS;