--- a/language-models/de/build.rs
+++ b/language-models/de/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    model_builder::build_frequencies("models");
+}