--- a/language-models/de/build.rs
+++ b/language-models/de/build.rs
@@ -1,3 +1,4 @@
 fn main() {
     model_builder::build_frequencies("models");
+    model_builder::generate_statics("../models", "src");
 }