--- a/language-models/eu/build.rs
+++ b/language-models/eu/build.rs
@@ -1,0 +1,4 @@
+fn main() {
+    model_builder::build_frequencies("models");
+    model_builder::generate_statics("../models", "src");
+}