--- a/language-models/it/build.rs
+++ b/language-models/it/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    model_builder::build_frequencies("models");
+}