--- a/language-models/lg/build.rs
+++ b/language-models/lg/build.rs
@@ -1,4 +1,4 @@
 fn main() {
     model_builder::build_frequencies("models");
-    model_builder::generate_statics("../models", "src");
+    model_builder::generate_statics("models", "src");
 }