As the slightly ironic title hopefully conveys, I'm not very convinced of this idea, I'm just putting it here for the sake of discussion.

# Motivation

Lingua stores language models (which really are just n-grams and their associated probabilities) as zipped JSON files in the binary. Depending on user preference either upfront at startup or lazily on demand, it will unzip, parse and then transform this data into the in-memory representation `HashMap<Ngram, f64>` used by part of the detection pipeline. That carries a certain computational cost, which is negligible for most use cases where the detector is reused and this initial cost can be offset over the detector's lifetime as all following detections can use the already "warmed up" detector.

In certain environments however, for example serverless computing or when using WebAssembly in plugins for other software, the detector can not be used in such a way. In these cases our code (containing a call to Lingua) is very short-lived and no reuse over subsequent invocations is possible, meaning each time it has to be constructed from scratch. In these permanent "cold start" situations, the described startup cost ends up dominating the overall runtime needed for detection. In my case, where I invoke Lingua from C# by means of [wasmtime](https://wasmtime.dev), that results in taking 540 ms for detecting the language of a short text, most of which is spent in startup. The following flamegraph illustrates this nicely. You can see some time being spent deflating the zipped data, some more deserializing the JSON and then some constructing the hash map.

![Flamegraph before changes](https://user-images.githubusercontent.com/5547518/160281152-cb3b4c30-ba41-4cbb-9d54-ee9a37397ee7.svg)

I would like to point out that this is not a weakness of Lingua. It's a problem that comes with the domain of language detection and I'm not aware of a library that has solved this. And as I mentioned, it's really only an issue in specialized use cases.

# Solution

## Idea

We could save a lot of time by embedding the language models into the binary in a way that is significantly cheaper to read. With [rkyv's zero-copy deserialization](https://rkyv.org/zero-copy-deserialization.html) it's possible to encode a data structure in a format that is the same as its in-memory representation. This means there's no additional work to be done, as language models can be read directly from the binary's data segment where they are stored.

## Implementation

I hacked the demo together by adding a build script for language model crates. It reads the JSON models for a language and writes them to binary files in their rkyv encoding. These bytes are then simply statically included and accessed at runtime. Instead of exposing a `Dir<'static>`, every language crate now has a `[Option<&'static [u8]>; 5]`, an array of bytes for all 5 possible n-gram types. The bytes are just the archived representation of `HashMap<Ngram, f64>`.

To read the JSON models from the build script I had to move the `Fraction` and `Ngram` types out into a separate `common` crate used by both the build script and Lingua.

## Results

I measured the C# scenario I described earlier.

- Benchmark implemented with BenchmarkDotNet on .NET 6.0
- Uses wasmtime to call a WebAssembly module built with lingua to detect the language of a single short text
- Instance of WebAssembly module (its memory space) is destroyed after each iteration, while the JIT-compiled module (bytecode) itself can be reused
- Only 4 languages were included in the binary

The effect of zero-copy deserialization of the language models in this cold start scenario is as follows:

|             | before | after  | change       |
|-------------|--------|--------|--------------|
| Binary size | 8.8 MB | 33 MB  | x3.75 bigger |
| Time        | 543 ms | 8.9 ms | x61 faster   |

The new flamegraph looks much less informative, mainly because there's just not a whole lot going on anymore. Some time is spent on what I think is language runtime stuff and the rest in the actual detection, there is nothing obvious to optimize away anymore.

![Flamegraph after changes](https://user-images.githubusercontent.com/5547518/160289053-1f7173ac-9801-4d4c-b9b5-2789837fbc45.svg)

These results are impressive both in the positive and negative sense. The build takes much longer now because it's serializing a bunch of pretty sizable files to disk for all languages. And the actual binary grows to a size that is no longer convenient to deliver in all environments, for example if you imagine it having to be downloaded over a mobile connection to a phone. On the other hand it's pretty amazing to see that you can initialize and run a sophisticated language detection suite in just under 9 ms.