@pemistahl 2022-03-28
Hi Martin, this is a very interesting case study. Thanks a lot for the effort you have put into it. It is surprising how much time is needed for unzipping and parsing, even though Rust already is very efficient performing this kind of task. But the binary size is problematic, I agree. Do you perhaps know an alternative to *rkyv* that produces smaller binaries but is still significantly faster than json decoding?

Can you do me a favor and make your PR compile properly? The CI pipeline has failed. I would then merge the PR into a feature branch of my own for future reference and experiments. Thank you.