@djkoloski 2022-03-28
If it matters, I was going to do some binary size investigation on this change to see if there were any space-saving opportunities that rkyv is missing. Do you have an idea of what an acceptable binary size threshold would be?