@martindisch 2022-03-28
@pemistahl I don't know of an alternative that ticks all the boxes. There are formats that are faster to deserialize and there are even faster JSON deserializers than `serde_json`, but I think it wouldn't make that much of a difference. Nothing comes close to zero-copy deserialization where absolutely no additional work has to be done. Also deserialization is only part of the equation, as we currently spend even slightly more time populating the n-gram hash map after deserializing the language model.

I fixed the remaining issues so the tests pass. But the pipeline takes over 25 minutes now and the one for Windows can't even complete because it runs out of disk space, I guess that speaks for itself :smile:

@djkoloski Great to see you here! By the way, I first learned about rkyv from https://github.com/swc-project/swc/issues/2635. Pretty amazing how this approach solved a major blocker and contributed to making the new plugin system feasible, that's a pretty big deal for SWC and the sizable (and only growing) ecosystem around it.

I can't give you a hard number for a size limit, but I don't think it matters either. Looking at the fivegrams for the German language model as an example, the intermediate file containing the serialized bytes takes up 6'732'788 bytes on disk for the total of 336'605 entries in the `HashMap<Ngram, f64>` (an `Ngram` is just a wrapper for a `String`). That means about 20 bytes per entry, of which the `f64` takes up 8 and the n-gram is a string of 5 characters, so at least 5 bytes unless they're not all ASCII and take up more. That's 13 bytes we're guaranteed to require for every entry and I imagine the rest is used to support the hash map itself. My napkin math may be wrong, but I'd be very impressed if there's much room for improvement here. There's just no way around it, at one point during the execution all this space for the uncompressed data will be taken up, it used to be just in memory during runtime, now that we're shifting the work to compile time it's going to be in the binary too. Through no fault of either Lingua or rkyv, there are just some theoretical limits we can't beat.