@djkoloski 2022-03-28
That's awesome! I've been idly watching swc for a while and it's been exciting to see everything they're working on. There's some new tech coming down the pipeline that I think will help make those kinds of use cases better too.

The archived hash map implementation for rkyv has an overhead of up to four bytes per key, so reducing that even by a byte or two could yield some big savings. Using a B-Tree might result in lower overhead too, at the expense of maybe slower lookup. An added bonus of using a B-Tree would also be faster serialization, since hash maps do a lot of heavy compute to build perfect hash tables. In situations like this, it might be worth resurrecting the older swisstable-based implementation since it has similar performance characteristics and better serialization speed.