@martindisch 2022-03-29
Nice, you're right a couple of bytes might actually make a difference here. This is turning out to be a pretty good benchmark for that sort of thing! I'm still going to close this PR because I just opened it to document the results of this little investigation. We can keep discussing here if anything comes up.