@Manishearth 2022-04-05
Worth noting: we've had a similar desire in ICU4X, but unfortunately we don't use `rkyv` (we instead use `serde` and `zerovec`), so there's no  way to to a zero-cost unvalidated deserialize of trusted data.

For this purpose, we designed [CrabBake](https://docs.google.com/document/d/192l7yr6hVnG11Dr8a7mDLonIb6c8rr6zq-iswrZtlXE/edit?usp=sharing), which allows you to serialize an object _straight to rust code_. We've got a working experimental implementation in [here](https://github.com/unicode-org/icu4x/tree/main/experimental/crabbake), hoping to have time to polish it up and start using it by the end of this month.