@martindisch 2022-04-06
Very interesting, thanks for putting it out there! This is exactly the kind of input I was hoping for when posting the writeup here.

It sounds like a much more elegant version of my first attempt before I switched over to `rkyv`. I started out by generating code that constructs a [PHF map](https://docs.rs/phf/0.10.1/phf/macro.phf_map.html) and then doing an `include!()` to pull in the code at compile time. I ended up abandoning this one because I didn't like how it had to generate enormous source files. That would be way more convenient with CrabBake.